+++ src/SignalBus.cpp
@@ -61,7 +61,7 @@ SignalBus::~SignalBus()
 
 void SignalBus::SetSignalCount( int signalCount )
 {
-    int fromSize = _signals.size();
+    int fromSize = (int)_signals.size();
 
     _signals.resize( signalCount );
 
@@ -73,7 +73,7 @@ void SignalBus::SetSignalCount( int signalCount )
 
 int SignalBus::GetSignalCount() const
 {
-    return _signals.size();
+    return (int)_signals.size();
 }
 
 Signal::SPtr const& SignalBus::GetSignal( int signalIndex ) const
